added functionsetting up ministep and function writing input to ministep
